✨ (src): add groups support (#24)
Added code to support groups key inside redis db. Now the dbms is able to return user's groups too.
User's groups are now used for admin authentication. If a user is in admin group will be displayed the admin-page.

## Commits

* ✨ (src/db_redis): add support for groups key

Added code to support `groups` key inside redis db.
Now the dbms is able to return user's groups too.

* ✨ (src/routes): get groups from redis

User's groups are now used for admin authentication.
If a user is in `admin` group will be displayed the admin-page.

* 🔨 (src/db_redis): add delete request check

Approved/synced requests cannot be deleted !
Added check to db side.

* 🔨 (src/templates/users): add delete check

Approved/synced requests cannot be removed !
If a request is `approved` the `delete` button is disabled.
If a request is `synced` `goto uninuvola` button is shown.

* ✏️ (src/templates/home): fix typo

Fix `logut`  ➡️ `logout`