🚑️ (src/db_redis): fix emty request check
Fixed a wrong if statement in `get_request_data()`: `request_data[key]` could
be `None` (because it's a string Key) or `set()` (because it's a set Key), respectively generated
by `__get_key()` and `__get_skey`. Need to check that req. data is not None and then is not an
empty set.