🚑️ (src/db_redis): fix empty request check (#30)
Fixed a wrong if statement in `get_request_data()`: `request_data[key]` could be `None` (because it's a string Key) or `set()` (because it's a set Key), respectively generated by `__get_key()` and `__get_skey()`. Need to check that req. data is not None and then is not an empty set.

Added ldapsync support and updated UI

## Commits

* 🚑️ (src/db_redis): fix emty request check

Fixed a wrong if statement in `get_request_data()`: `request_data[key]` could
be `None` (because it's a string Key) or `set()` (because it's a set Key), respectively generated
by `__get_key()` and `__get_skey`. Need to check that req. data is not None and then is not an
empty set.

* 🚑 (src/routes): fix `KeyError` for new requests

* 🔥 (src/): remove `ADMIN_USERS`

Removed old `ADMIN_USERS` config from code.

* 📝 (README): remove `ADMIN_USERS`

* ✨ (src/db_redis): update groups when updating request

When a requests is approved (passing from `pending` to `approved`),
`groups` key are set by adding `users`.

* ✨ (src/): add user infos

Added functions to get and use redis userinfos keys.

* 🐛 (src/db_redis): fix wrong var name

* 🐛 (src/db_redis): fix missing string in logging

* 💄 (src/templates/users/admin): add full user's name

* 🐛 (src/routes): fix wrong argument

* 🐛 (src/db_redis): fix wrong redis key format

* 💄 (src/): add user's name in admin view

* ✨ (src/): add notify ldapsync function

* 🐛 (src/db_redis): add protocol to ldap notify

* 💄 (src/templates/users/user): update user's name